requirements.txt
streamlit
langchain
langchain-groq
transformers
chromadb
pypdf
python-dotenv
sentence-transformers